cleanup, fix(exec): I cleaned and improved some stuff there and there, I also fixed a bad issue that printed the timestamp in the wrong way (lag of 1ms from time to time) at the cost of performance but the project still work on the older computer at school so all good for now.
