+++ philo/srcs/exec/philo.c
@@ -3,108 +3,46 @@
 #include "exec.h"
 #include <unistd.h> //usleep
 
-void routine_take_fork(t_philo *philo, bool right)
+static int create_thread(t_philo *philo, t_philo *tmp, long now, bool even)
 {
-        if(right)
-        {
-            pthread_mutex_lock(&philo->right->mutex);
-            philo->right->available = false;
-            print_msg_routine(philo, IS_TAKING_FORK);
-        }
-        else
-        {
-            pthread_mutex_lock(&philo->left->mutex);
-            philo->left->available = false;
-            print_msg_routine(philo, IS_TAKING_FORK);
-        }
-}
-void* routine_odd(void *arg)
-{
-    t_philo *philo = (t_philo *)arg;
-    //struct timeval tv;
-    //TODO mettre le t_eat /2 directement dans la struct pour eviter calcul redondant
-    usleep(1000);//moitie wait la moitie de manger pour creer un decalage
-    pthread_mutex_lock(&philo->set->death_mutex);
-    while(!philo->set->death && philo->meals_eaten != philo->set->max_meal)
+    tmp = philo;
+    while(philo)
     {
-        pthread_mutex_unlock(&philo->set->death_mutex);
-        print_msg_routine(philo, IS_THINKING);
-        usleep(250);
-        routine_take_fork(philo, true);
-        routine_take_fork(philo, false);
-        print_msg_routine(philo, IS_EATING);
-        //usleep_routine(philo->set->t_eat, philo);
-        if(philo->set->t_eat > philo->set->t_die)
-            usleep(philo->set->t_die);
+        now = fill_now();
+        pthread_mutex_lock(&philo->t_alive_mutex);
+        philo->t_alive = now;
+        pthread_mutex_unlock(&philo->t_alive_mutex);
+        if(even)
+        {
+            if(pthread_create(&philo->thread_id, NULL, &routine_even, philo) != 0)
+                return(RETURN_FAILURE);
+            even = false;
+        }
         else
-            usleep(philo->set->t_eat);
-        philo->left->available = true;
-        pthread_mutex_unlock(&philo->left->mutex);
-        philo->right->available = true;
-        pthread_mutex_unlock(&philo->right->mutex);
-        print_msg_routine(philo, IS_SLEEPING);
-        if(philo->set->t_eat + philo->set->t_sleep > philo->set->t_die)
-            usleep(philo->set->t_die);
-        else
-        usleep(philo->set->t_eat);
-        //usleep_routine(philo->set->t_sleep, philo);
-        //usleep(philo->set->t_sleep);   
-        pthread_mutex_lock(&philo->set->death_mutex);
+        {
+            if (pthread_create(&philo->thread_id, NULL, &routine_odd, philo) != 0)
+                return(RETURN_FAILURE);
+            even = true;
+        }
+        philo = philo->next;
     }
-    pthread_mutex_unlock(&philo->set->death_mutex);
-    pthread_mutex_lock(&philo->set->pasta_mutex); 
-    if(philo->meals_eaten == philo->set->max_meal)
-        philo->set->philo_full_pasta = philo->set->philo_full_pasta + 1;
-    pthread_mutex_unlock(&philo->set->pasta_mutex); 
-    return(0);
+    if(pthread_create(&tmp->set->monitor_thread_id, NULL, &philo_monitor, tmp) != 0)
+        return(RETURN_FAILURE);
+    return(RETURN_SUCCESS);
 }
 
-void* routine_even(void *arg)
+static int join_thread(t_philo *philo)
 {
-    t_philo *philo = (t_philo *)arg;
-    pthread_mutex_lock(&philo->set->death_mutex);
-    while(!philo->set->death && philo->meals_eaten != philo->set->max_meal)
+
+    if(pthread_join(philo->set->monitor_thread_id, NULL))
+        return(RETURN_FAILURE); //protect
+    while(philo)
     {
-        pthread_mutex_unlock(&philo->set->death_mutex);
-        print_msg_routine(philo, IS_THINKING);
-        usleep(250);
-        routine_take_fork(philo, false);
-        if(philo->right->id == philo->left->id)
-        {
-            pthread_mutex_lock(&philo->set->death_mutex);
-            pthread_mutex_unlock(&philo->left->mutex); 
-            break;
-            // while(!philo->set->death)
-            //     print_msg_routine(philo, WAIT);
-            
-        }
-        routine_take_fork(philo, true);
-        print_msg_routine(philo, IS_EATING);
-        if(philo->set->t_eat > philo->set->t_die)
-            usleep(philo->set->t_die);
-        else
-        usleep(philo->set->t_eat);
-        //usleep_routine(philo->set->t_eat, philo);
-        //usleep(philo->set->t_eat);
-        philo->left->available = true;
-        pthread_mutex_unlock(&philo->left->mutex);
-        philo->right->available = true;
-        pthread_mutex_unlock(&philo->right->mutex);
-        print_msg_routine(philo, IS_SLEEPING);
-        if(philo->set->t_eat + philo->set->t_sleep > philo->set->t_die)
-            usleep(philo->set->t_die);
-        else
-        usleep(philo->set->t_eat);
-        //usleep_routine(philo->set->t_sleep, philo);
-        //usleep(philo->set->t_sleep);  
-        pthread_mutex_lock(&philo->set->death_mutex);  
+        if(pthread_join(philo->thread_id, NULL))
+            return(RETURN_FAILURE); //protect?
+        philo = philo->next;
     }
-    pthread_mutex_unlock(&philo->set->death_mutex);
-    pthread_mutex_lock(&philo->set->pasta_mutex); 
-    if(philo->meals_eaten == philo->set->max_meal)
-        philo->set->philo_full_pasta = philo->set->philo_full_pasta + 1;
-    pthread_mutex_unlock(&philo->set->pasta_mutex);
-    return(0);
+    return(RETURN_SUCCESS);
 }
 
 void philosopher(t_settings *set, t_philo *philo, t_forks *forks)
